[TLERaw][Mega] Update multi-rank candidate to v234
