+++ python/test/tle/integration/megamoe/RUNBOOK_CN.md
@@ -1,69 +1,97 @@
-# Triton TLE MegaMoE 已跑通 Case
+# Triton TLE MegaMoE 已验证入口
 
-本文档是当前目录唯一说明文档，只记录这个隔离 MegaMoE operator 已跑通的 case 和手动复现命令。
+本文只记录这个隔离 MegaMoE operator 的当前入口、依赖、验证证据和手动复现命令。
 
-## 当前 production-shape 候选（2026-08-03）
+## 当前 production-shape 候选（2026-09-01）
 
-当前单卡与多卡候选已更新为：
-
-| 场景 | runner | 结构 | 当前状态 |
+| 场景 | runner | 定位 | 当前状态 |
 |---|---|---|---|
-| 单卡 | `megamoe_operator/production/v25/run.py` | BM64、单 math warp-group、D8 descriptorless raw TMA1D、v23 wide L2 scatter | NP1 MoE-7 correctness PASS |
-| 8 rank | `megamoe_operator/production/v33/run.py` | BM128、双 math warp-group、SMEM expert count、并行 NVLink signal、单 D8 TMA1D stream | MoE-7 8/8 correctness PASS |
+| 单 rank 回归 | `megamoe_operator/production/v25/run.py` | BM64、单 math warp-group、D8 descriptorless TMA1D、wide L2 scatter | NP1 MoE-7 correctness PASS |
+| 多 rank 主线 | `megamoe_operator/production/v234/run.py` | BM128、双 math warp-group、双 D8 pull stream、独立 A/SFA 与 B TMA producer、真实 W1 布局 | H100 MoE-7 np8/t512 8/8 PASS |
 
-两者都是一个 rank 一次 Triton/TLE persistent-kernel launch。raw CUDA 边界只承载
-D8 TMA1D 和 L2 scatter；routing、TLE pipe、WGMMA、SwiGLU、scale 和 combine
-仍在 Triton/TLE kernel 中。
+v234 是 `zhiyuan_megakernel` 当前 manifest 中的 multi-rank mainline，来源提交为
+`fdb7f8e`，打包前 immutable kernel SHA256 为
+`d57fb2252c63818a0058f936ff3ed46b9e7fadba29858f9f47eadb9526d6464b`。
+本目录只调整运行时和 raw helper 的相对路径，不改变 kernel 语义。
 
-### 编译器基线
+每个 rank 只发射一次 Triton/TLE persistent kernel。Triton/TLE 保留 pipe、
+WGMMA、scale、SwiGLU、L1/L2 和 combine；raw CUDA 边界仅承载 D6-D9 的
+descriptorless TMA1D dispatch/pull 以及 L2 remote scatter。
 
-这两个 production-shape runner 需要新的 pipe API，以及 warp-specialized
-helper 中 `tle.local_pointers` 的安全 inlining。PR837 的第一版实现曾被 PR845
-回退；更窄的 TLE-specific 实现随后由 PR859 重新合入 `main`。本分支因此直接
-forward-port 到包含 PR859 的 upstream `main`，不携带已回退的 PR837 实现。
+## 编译器依赖
 
-### 统一 workload
+v25 依赖当前 main 已有的 pipe/warp-specialization 支持。v234 还依赖经过
+验证的 PR837 TLE compiler build；算子实际走到的新增能力为：
+
+- `buffered_tensor.subslice`：保持 rank、SMEM layout 和 allocation shape；
+- 同一 pipe 上多个纯 TMA writer；
+- 多 writer 推导出的 TMA token `full_count` lowering。
+
+v234 会在导入 Triton 前设置 `TLE_MULTI_TMA_WRITERS=1`。未合入上述编译器
+能力的 FlagTree 不能编译该入口。
+
+当前已验证的组合是本 PR 中的 v234 加历史 PR837 build：H100 NP2 synthetic
+correctness 为 2/2 rank PASS，历史 H100 NP8 真实数据为 8/8 rank PASS。另行在
+FlagTree current main 上只叠加 `subslice` 与 multi-writer 两个精简提交的组合，
+虽然能够完成编译和 launch，但 NP2 在运行时触发 CUDA illegal instruction；该
+组合不能视为已支持，也不能替代完整的配套 compiler 验证。
+
+## 统一 workload
 
 ```text
 H=4096, I=1536, E=128, topk=8, tokens/rank=512, drop=0, stages=4
 ```
 
-单卡 v25 correctness：
-
-```bash
-MEGAMOE_NP=1 W_NTOK=512 W_TOPK=8 W_NEXP=128 W_K=4096 W_INTER=1536 \
-W_DROP=0 W_STAGES=4 W_BENCH=0 W_D8_TMA1D_LEVEL=2 \
-python python/test/tle/integration/megamoe/megamoe_operator/production/v25/run.py
-```
-
-8-rank v33 correctness：
+### 8-rank synthetic correctness
 
 ```bash
 MEGAMOE_NP=8 W_NTOK=512 W_TOPK=8 W_NEXP=128 W_K=4096 W_INTER=1536 \
-W_DROP=0 W_STAGES=4 W_BENCH=0 W_D8_TMA1D_LEVEL=2 \
-W_SMEM_EXPERT_COUNT=1 W_FAST_NVLINK_BARRIER=1 W_D8_PULL_STREAMS=1 \
-python python/test/tle/integration/megamoe/megamoe_operator/production/v33/run.py
+W_DROP=0 W_STAGES=4 W_BENCH=0 W_TIMEOUT=900 \
+python python/test/tle/integration/megamoe/megamoe_operator/production/v234/run.py
 ```
 
-benchmark 在上述命令中改为：
+通过标准为 8 个 rank 都输出：
 
 ```text
-W_BENCH=1 W_WARMUP=10 W_ITERS=30 W_BENCH_REDUCE=mean
+partials=4096 scatter_bad=0 errors=0 ... -> PASS
 ```
 
-本次 forward-port 后复测：
+### 真实 Qwen3 FP8 数据
 
-| candidate | correctness | event latency |
-|---|---:|---:|
-| v25 / NP1 | 1/1 PASS；4,096 partial rows；`scatter_bad=0 errors=0` | 1,272.5 us（PR837-compatible validation build） |
-| v33 / NP8 | 8/8 PASS；每 rank 4,096 partial rows；`scatter_bad=0 errors=0` | rank mean 473.54 us，rank std 4.70 us |
+设置 `MEGAMOE_SHARED_DATA_DIR` 后，v234 会读取 CUDA/TLE 共用的数据集，并在
+host preprocessing 中完成 checkpoint W1 的 SM90 gran-8 gate/up 交织：
 
-参考 CUDA event baseline 为 335.8 us，故本次 v33 复测为 CUDA 的约 70.9%。
+```bash
+MEGAMOE_SHARED_DATA_DIR=/path/to/qwen3_fp8_shared \
+MEGAMOE_NP=8 W_NTOK=512 W_TOPK=8 W_NEXP=128 W_K=4096 W_INTER=1536 \
+W_DROP=0 W_STAGES=4 W_BENCH=0 W_TIMEOUT=900 \
+python python/test/tle/integration/megamoe/megamoe_operator/production/v234/run.py
+```
 
-### 等价性边界
+已归档证据为 H100 np8/t512 8/8 rank PASS、`scatter_bad=0 errors=0`；独立
+PyTorch oracle 的 relative L2 为 `0.0032209039`，cosine 为
+`0.9999948372`。
 
-| CUDA behavior | 当前 TLE behavior | 状态 | 未闭合项 |
+### Event benchmark
+
+在 correctness 通过后使用同一 workload：
+
+```bash
+MEGAMOE_NP=8 W_NTOK=512 W_TOPK=8 W_NEXP=128 W_K=4096 W_INTER=1536 \
+W_DROP=0 W_STAGES=4 W_BENCH=1 W_WARMUP=10 W_ITERS=30 \
+W_BENCH_REDUCE=mean W_GPU_START_BARRIER=1 W_TIMEOUT=900 \
+python python/test/tle/integration/megamoe/megamoe_operator/production/v234/run.py
+```
+
+不同 revision、输入、时钟状态或 rank reduction 得到的 latency 不可直接与
+CUDA 数字混算。本 PR 不新增跨实现性能比例声明。
+
+## UserHopper 等价性边界
+
+| CUDA behavior | v234 behavior | 状态 | 仍未闭合项 |
 |---|---|---|---|
-| D8 FP8 payload TMA pull | descriptorless TMA1D；SF/top-k 保持普通 load | `partial` | stage-specific lifecycle evidence 尚未覆盖完整 CUDA contract |
-| L1/L2 WGMMA 与 combine | 单 kernel 内完成并通过当前 exact/数值检查 | `partial` | workspace cleanup/reuse 和全部 UserHopper stage parity 未闭合 |
-| 双 math warp-group | v33 BM128 两个 math WG | `partial` | 双 D8 pull stream 仍被 layout lowering 阻塞，默认保持 1 stream |
+| SM90 W1 FP8 transform | host 固化 `[all gate | all up]` 到 gran-8 gate/up 交织 | `equivalent`（当前 H100 np8/t512 scope） | 其它模型层和 shape 尚未覆盖 |
+| W1/L2 weight SF layout | W1 使用 `[EPR, NL1N, NK1]`，gate/up 六处地址已修复；L2 SF 保持 CUDA contract | `equivalent`（当前测试 scope） | 其它模型层和 shape 尚未覆盖 |
+| D8 FP8 payload pull | 双 stream descriptorless TMA1D；activation SF/top-k weight 保持普通 load | `partial` | 完整 D0-D10 stage JSON 和 workspace lifecycle 尚未闭合 |
+| TMA/WGMMA/WS performance path | BM128、双 math WG、独立 A/SFA 与 B TMA producer | `partial` | 仍需相同输入和相同计时口径下的 UserHopper matched 复测 |
+| combine/cleanup | 单 kernel 内完成当前 combine 与输出校验 | `partial` | UserHopper TMA/SMEM/mbarrier combiner 和完整 epoch/reuse contract 尚未闭合 |
