+++ python/test/tle/integration/megamoe/megamoe_operator/production/v234/__init__.py
@@ -1,0 +1,1 @@
+"""Validated v234 multi-rank MegaMoE operator."""
