+++ python/test/tle/integration/megamoe/megamoe_operator/production/v234/raw_l2_tma_scatter.cu
@@ -6,21 +6,19 @@
 // global address. That is required here because every row may target a
 // different rank/token/top-k slot. The row pointers are resolved by Triton and
 // staged in shared memory before this boundary.
-extern "C" __device__ void TleL2TmaScatter(
-    const uint64_t values_smem,
-    const uint64_t dst_rows_smem,
-    const int valid_rows,
-    const int cols,
-    const int bar_id) {
+extern "C" __device__ void TleL2TmaScatter(const uint64_t values_smem,
+                                           const uint64_t dst_rows_smem,
+                                           const int valid_rows, const int cols,
+                                           const int bar_id) {
   const int lane = threadIdx.x & 31;
   const int warp_in_wg = (threadIdx.x >> 5) & 3;
   const int row_bytes = cols * 2;
 
-  asm volatile("bar.sync %0, 128;" :: "r"(bar_id) : "memory");
+  asm volatile("bar.sync %0, 128;" ::"r"(bar_id) : "memory");
   asm volatile("fence.proxy.async.shared::cta;" ::: "memory");
 
   if (lane == 0) {
-    #pragma unroll 1
+#pragma unroll 1
     for (int j = 0; j < 16; ++j) {
       const int row = warp_in_wg * 16 + j;
       if (row < valid_rows) {
@@ -31,15 +29,14 @@ extern "C" __device__ void TleL2TmaScatter(
         uint64_t src_shared;
         uint64_t dst_row_shared;
         uint64_t dst;
-        asm volatile(
-            "cvta.to.shared.u64 %0, %3;\n"
-            "cvta.to.shared.u64 %1, %4;\n"
-            "ld.shared.u64 %2, [%1];\n"
-            "cp.async.bulk.global.shared::cta.bulk_group "
-            "[%2], [%0], %5;"
-            : "=l"(src_shared), "=l"(dst_row_shared), "=l"(dst)
-            : "l"(src_generic), "l"(dst_row_generic), "r"(row_bytes)
-            : "memory");
+        asm volatile("cvta.to.shared.u64 %0, %3;\n"
+                     "cvta.to.shared.u64 %1, %4;\n"
+                     "ld.shared.u64 %2, [%1];\n"
+                     "cp.async.bulk.global.shared::cta.bulk_group "
+                     "[%2], [%0], %5;"
+                     : "=l"(src_shared), "=l"(dst_row_shared), "=l"(dst)
+                     : "l"(src_generic), "l"(dst_row_generic), "r"(row_bytes)
+                     : "memory");
       }
 
       // Two groups of at most eight row copies per issuing warp. Waiting here
@@ -51,5 +48,5 @@ extern "C" __device__ void TleL2TmaScatter(
     }
   }
 
-  asm volatile("bar.sync %0, 128;" :: "r"(bar_id) : "memory");
+  asm volatile("bar.sync %0, 128;" ::"r"(bar_id) : "memory");
 }
