--- python/test/tle/integration/megamoe/megamoe_operator/production/v33/__init__.py
@@ -1,1 +1,0 @@
-"""V33 multi-rank candidate."""
